Move event handler definitions over to model/eventHandlers
-  Misplaced under `events`
-  Events dir now contains just the event definitions

JIRA: RAIL-3471
