+++ libs/sdk-ui-dashboard/src/plugins/exampleDashboardComponent.tsx
@@ -5,7 +5,7 @@ import React from "react";
 import { idRef, IInsight, insightTags, insightTitle } from "@gooddata/sdk-model";
 import { IInsightWidget } from "@gooddata/sdk-backend-spi";
 import includes from "lodash/includes";
-import { anyDashboardEventHandler, singleEventTypeHandler } from "../model/events/eventHandler";
+import { anyDashboardEventHandler, singleEventTypeHandler } from "../model/eventHandlers/eventHandler";
 
 //
 //
